requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
cryptography==44.0.2
    # via -r requirements.in
deprecated==1.2.18
    # via limits
distlib==0.3.9
    # via virtualenv
filelock==3.18.0
    # via virtualenv
flask==2.3.3
    # via
    #   -r requirements.in
    #   flask-limiter
    #   flask-sqlalchemy
flask-limiter==3.12
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
identify==2.6.9
    # via pre-commit
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
limits==4.6
    # via flask-limiter
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.21.1
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
nodeenv==1.9.1
    # via pre-commit
ordered-set==4.1.0
    # via flask-limiter
packaging==24.2
    # via
    #   limits
    #   marshmallow
platformdirs==4.3.7
    # via virtualenv
pre-commit==3.6.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.22
    # via cffi
pygments==2.19.1
    # via rich
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via pre-commit
requests==2.31.0
    # via -r requirements.in
rich==13.9.4
    # via flask-limiter
sentry-sdk==2.25.1
    # via -r requirements.in
sqlalchemy==2.0.40
    # via flask-sqlalchemy
tenacity==8.2.2
    # via -r requirements.in
typing-extensions==4.13.1
    # via
    #   limits
    #   sqlalchemy
urllib3==2.3.0
    # via
    #   requests
    #   sentry-sdk
virtualenv==20.30.0
    # via pre-commit
werkzeug==3.1.3
    # via flask
wrapt==1.17.2
    # via deprecated
